Adding Function to calc on controller and view to display calculated data

+++ Models/Product.cs
@@ -2,6 +2,6 @@ namespace SaloonHelper.Models
 {
     public class Product
     {
-        public int InitialPrice { get; set; }
+        public double InitialPrice { get; set; }
     }
 }
